refactor: extract image utilities to service
